+++ frontend/src/Actions/AttachForm.svelte
@@ -1,9 +1,10 @@
 <script lang="ts">
   import { Form, InputGroup, InputGroupText } from "@sveltestrap/sveltestrap";
   import DateTime from "../Widgets/DateTime.svelte";
-  import { attachments, types, parts } from "../lib/store";
-  import { AttEvent, maxDate, Part } from "../lib/types";
+  import { attachments, types } from "../lib/store";
+  import { AttEvent, maxDate } from "../lib/types";
   import { by, filterValues } from "../lib/mapable";
+  import { parts, Part } from "../Part/part";
 
   function lastDetach(part: Part) {
     let last = filterValues($attachments, (a) => a.part_id == part.id).sort(
