+++ frontend/src/Actions/AttachPart.svelte
@@ -2,8 +2,9 @@
   import { Modal, ModalHeader, ModalBody } from "@sveltestrap/sveltestrap";
   import ModalFooter from "./ModalFooter.svelte";
   import { types } from "../lib/store";
-  import { AttEvent, Part } from "../lib/types";
+  import { AttEvent } from "../lib/types";
   import AttachForm from "./AttachForm.svelte";
+  import { Part } from "../Part/part";
 
   let attach: AttEvent;
   let part: Part | undefined;
