+++ frontend/src/Actions/ReplacePart.svelte
@@ -5,11 +5,12 @@
     ModalHeader,
     ModalBody,
   } from "@sveltestrap/sveltestrap";
-  import { AttEvent, Attachment, Part, Type, maxDate } from "../lib/types";
-  import { types, parts, user } from "../lib/store";
+  import { AttEvent, Attachment, Type, maxDate } from "../lib/types";
+  import { types, user } from "../lib/store";
   import ModalFooter from "./ModalFooter.svelte";
   import NewForm from "./NewForm.svelte";
   import Dispose from "../Widgets/Dispose.svelte";
+  import { parts, Part } from "../Part/part";
 
   let part: Part, oldpart: Part, newpart: Part;
   let type: Type;
