+++ frontend/src/Part/PartHist.svelte
@@ -1,8 +1,9 @@
 <script lang="ts">
   import { filterValues, by } from "../lib/mapable";
-  import { parts, attachments, types } from "../lib/store";
+  import { attachments, types } from "../lib/store";
   import Usage from "../Usage/Usage.svelte";
   import { link } from "svelte-spa-router";
+  import { parts } from "./part";
 
   export let id: number;
