+++ frontend/src/Part/Subparts.svelte
@@ -1,11 +1,12 @@
 <script lang="ts">
-  import { Attachment, Type, Part } from "../lib/types";
+  import { Attachment, Type } from "../lib/types";
 
   import { Table } from "@sveltestrap/sveltestrap";
   import { filterValues, by } from "../lib/mapable";
   import { attachments, types } from "../lib/store";
   import SubType from "./SubType.svelte";
   import Wizard from "./Wizard.svelte";
+  import { Part } from "./part";
 
   export let gear: Part;
   export let hook: Type;
