+++ frontend/src/Service/NewService.svelte
@@ -7,7 +7,7 @@
   } from "@sveltestrap/sveltestrap";
   import ModalFooter from "../Actions/ModalFooter.svelte";
   import { Service } from "./service";
-  import type { Part } from "../lib/types";
+  import { Part } from "../Part/part";
   import ServiceForm from "./ServiceForm.svelte";
 
   let part: Part;
